server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Servidor do Fast Entregas"
authors = ["tuliomitico <tuliofc2008@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
MarkupSafe = "2.0.1"
Flask = "1.1.4"
Flask-Cors = "3.0.8"
ariadne = "0.14.0"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
SQLAlchemy = "~1.4.22"
Flask-SQLAlchemy = "^2.5.1"
alembic = "^1.7.7"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.0"
importlib-resources = "^5.7.1"
importlib-metadata = "^4.11.3"
SQLAlchemy-Utils = "^0.38.2"
PyJWT = "^2.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
marshmallow-enum = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
factory-boy = "^3.2.1"

[tool.pytest.ini_options]
addopts = "-vv -cov"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
